Migration to gwt-uibinder-example for testbed This change used the gwt-maven-archetype to generate client/shared/server modules in a way that works better with gwt app compilation.
`gwt-uibinder-example` module (parent of *-{shared,client,server}) does not inherit from the root maven module.  This was intentional as this example is geared towards a "real world" scenario.
This means that some maven properties, plugins, etc will need to be updated as necessary as they will not inherit from the root module.